requirements.txt
anyio==3.7.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.7.2
astroid==2.15.6
asttokens==2.2.1
async-lru==2.0.3
Babel==2.12.1
backcall==0.2.0
black==23.7.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.3
dateparser==1.1.8
defusedxml==0.7.1
Deprecated==1.2.14
dj-rest-auth==5.0.2
Django==4.2.4
django-cors-headers==4.3.0
django-debug-toolbar==4.2.0
django-extensions==3.2.3
django-filter==23.3
django-oauth-toolkit==2.3.0
django-oauth2-provider==0.2.6.1
django-reset-migrations==0.4.0
django-shortcuts==1.6
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-extra-fields==3.7.0
drf-social-oauth2==2.1.3
drf-yasg==1.21.7
ecdsa==0.18.0
executing==2.0.0
Faker==12.0.1
filetype==1.2.0
flake8==6.1.0
flake8-broken-line==1.0.0
flake8-commas==2.1.0
flake8-django==1.4
flake8-isort==6.0.0
flake8-plugin-utils==1.3.3
flake8-print==5.0.0
flake8-return==1.2.0
gunicorn==21.2.0
icecream==2.1.3
idna==3.4
inflection==0.5.1
isort==5.12.0
jwcrypto==1.5.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
mixer==7.2.2
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.1
pathspec==0.11.2
petname==2.6
Pillow==10.0.0
platformdirs==3.10.0
psycopg2==2.9.7
pyasn1==0.5.0
pycodestyle==2.11.0
pycparser==2.21
pyflakes==3.1.0
Pygments==2.16.1
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
shortuuid==1.0.11
six==1.16.0
sniffio==1.3.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.0.4
wrapt==1.15.0
